requirements.txt
flask==2.0.1
flask-socketio==5.1.1
psycopg2-binary==2.9.1
scikit-learn==0.24.2
pandas==1.3.0
python-dotenv==0.19.0